Refactor ban cmd
This in theory should work, but I can't test it because I'm missing your XP library and can't remember how to get it :p
