+++ commands/lyrautilities/ban/messages/ban.message.js
@@ -1,0 +1,8 @@
+const {baseEmbed} = LyraCore;
+
+module.exports = function (userToBan, reason) {
+    return baseEmbed()
+        .addField('Banned:', userToBan)
+        .addField('Reason', reason)
+        .setImage('https://tenor.com/view/when-your-team-too-good-ban-salt-bae-gif-7580925')
+}
